+++ src/graphics/texmap_projection.cpp
@@ -69,7 +69,7 @@ namespace bricksim::graphics::texmap_projection {
                 polyWrapper.push_back(poly2d);
                 const auto triangleIndices = mapbox::earcut(polyWrapper);
 
-                for (int i = 0; i < triangleIndices.size(); i += 3) {
+                for (size_t i = 0; i < triangleIndices.size(); i += 3) {
                     std::array<glm::vec2, 3> UVs{};
                     std::array<glm::vec3, 3> coords3d{};
                     bool allUVsInImage = true;
