+++ src/helpers/stringutil.cpp
@@ -275,7 +275,7 @@ namespace bricksim::stringutil {
     std::string repeat(const std::string& str, unsigned int times) {
         std::string result;
         result.reserve(str.size() * times + 1);
-        for (int i = 0; i < times; ++i) {
+        for (size_t i = 0; i < times; ++i) {
             result += str;
         }
         return result;
